Refactor admin.controller.ts to improve admin login functionality and handle admin creation and deletion

+++ src/controllers/Admin/admin.controller.ts
@@ -2,7 +2,6 @@ import { Request, Response } from 'express';
 import AdminService from '../../services/AdminServices/admin.service';
 import { AdminAuthenticatedRequest } from '../../middlewares/authMiddleware';
 import { BadRequestError, ForbiddenError } from '../../utils/customErrors';
-import { ADMIN_EMAIL } from '../../utils/constants';
 import { AuthUtil } from '../../utils/token';
 import { emailService, EmailTemplate } from '../../utils/Email';
 
@@ -11,15 +10,10 @@ export default class AdminController {
     static async loginSuperAdmin(req: Request, res: Response) {
         const { email } = req.body;
 
-        let emailToUse = email.toLowerCase().trim();
-        let firstName = 'Owner';
-        if (email !== ADMIN_EMAIL) {
-            const checkAdmin = await AdminService.getAdminByEmail(email);
-            emailToUse = checkAdmin.email;
-            firstName = checkAdmin.name.split(' ')[0];
-        }
+        const checkAdmin = await AdminService.getAdminByEmail(email);
+        const firstName = checkAdmin.name.split(' ')[0];
 
-        const otpCode = await AuthUtil.generateCode({ type: 'adminlogin', identifier: emailToUse, expiry: 60 * 10 });
+        const otpCode = await AuthUtil.generateCode({ type: 'adminlogin', identifier: checkAdmin.email, expiry: 60 * 10 });
 
         const templateData = {
             otpCode,
@@ -28,7 +22,7 @@ export default class AdminController {
         
         // Send email with OTP
         await emailService.send({
-            email: emailToUse,
+            email: checkAdmin.email,
             from: 'auth',
             subject: 'Admin Login Verification',
             html: await new EmailTemplate().adminLogin({ otpCode, name: firstName }),
@@ -49,26 +43,20 @@ export default class AdminController {
     static async verifySuperAdminLogin(req: Request, res: Response) {
         const { email, otpCode } = req.body;
 
-        let emailToUse = email.toLowerCase().trim();
-        let adminData = { email: emailToUse, name: 'Owner', isSuperAdmin: true };
-        if (email !== ADMIN_EMAIL) {
-            const checkAdmin = await AdminService.getAdminByEmail(email);
-            emailToUse = checkAdmin.email;
-            adminData = checkAdmin;
-        }
+        const checkAdmin = await AdminService.getAdminByEmail(email);
 
-        const validCode = await AuthUtil.compareAdminCode({ identifier: emailToUse, tokenType: 'adminlogin', token: otpCode });
+        const validCode = await AuthUtil.compareAdminCode({ identifier: checkAdmin.email, tokenType: 'adminlogin', token: otpCode });
         if (!validCode) {
             throw new BadRequestError('Invalid verification code');
         }
 
         // Generate admin token
-        const adminToken = await AuthUtil.generateAdminToken({ type: 'admin', identifier: emailToUse });
+        const adminToken = await AuthUtil.generateAdminToken({ type: 'admin', identifier: checkAdmin.email });
 
         res.status(200).json({
             status: 'success',
             message: 'Admin login successful',
-            data: { adminToken, admin: adminData },
+            data: { adminToken, admin: checkAdmin },
         });
     }
 
@@ -77,10 +65,6 @@ export default class AdminController {
 
         if (req.isSuperAdmin === false) {
             throw new ForbiddenError('Only super admin can create new admins');
-        }
-
-        if (email === ADMIN_EMAIL) {
-            throw new BadRequestError('Admin with this email already exists');
         }
 
         const newAdmin = await AdminService.createAdmin({ name, email, isSuperAdmin });
@@ -112,6 +96,11 @@ export default class AdminController {
 
         if (req.isSuperAdmin === false) {
             throw new ForbiddenError('Only super admin can delete admins');
+        }
+
+        // admin cannot delete self
+        if (id === req.admin.id) {
+            throw new ForbiddenError('You cannot delete yourself');
         }
 
         await AdminService.deleteAdmin(id as string);
